// src/app/models/Empaque.ts
export interface Empaque {
    idpacking?: number;
    packingName?: string;
    packingCode?: string;
    packingDescription?: string;
    image1?: string;
    image2?: string;
    image3?: string;
    packingType?: string;
    packingTypeText?: string;
    packingStateText?: string;
    packingState?: number;
    packing_idpacking?: number;
}